Implement audio packaging and export (#11)
Enables packaging the encoded audio from multiple files with metadata (sequence.json) files. The file system operations and further checks on the metadata are implemented in a new export-server.  The user interface now has an export button on the review page that offers a choice of export type - only the "audio only" type is implemented so far. The UI also shows progress and offers a Save-As dialog when finished.

* Describe methods to be implemented in the export-server.

* Add export server to electron-app's built-in API server.

* Implement user interface components for export process.

The user has to choose what kind of export they would like, and will
then be informed about the progress and any errors. The actions are
not yet implemented, beyond updating the UI.

* Describe task results in export server

* Add fs-extra dependency

It wraps the native fs module so its function can be used as promises,
and provides extras such as remove to rm -rf.

* Implement export worker for audio (untested)

* Remove duplicate file added by mistake.

* Create ExportService and connect UI to export server

The ExportService is akin to the FileService, it communicates with the
export server.

The export actions use the ProjectStore to get the sequence information.

So far, this can create the temporary export folder and display its path
on the file system.

* rebuild server code before launching electron app.

* Support sequence duration and loop flag.

* Allow saving exports to a specific folder.

Opens a Save As dialog when the export is finished and moves temporary
export folder to that location.

* Parse mdoObjects settings as integers, export sequence.json with white space

* Include sequenceId in baseURL in DASH manifests.

This is required because bbcat-js cannot understand relative paths.
The sequenceId has to be passed into the encoding workers, and the
audio/ prefix must be used here and in the export worker to build
the paths to sequence.json, DASH manifests, and segments within.

* add concurrently to run frontend as part of electron dev task

* fix re-encoding on re-open: s/basePath/encodedItemsBasePath/

This bug was due to the previously encoded items using the
encodedItemsBasePath name, which is checked to test whether they need to
be encoded again, and the encoding results using just basePath.
When the combined results were stored, the encodedItemsBasePath was
reset to undefined for previously encoded files, and so they were
assumed to be missing the next time the project was opened.

* Disable export methods not implemented.

* Add method to open the export in Finder.

* Include non-main sequences in audio export.

* remove a few debug prints.

* move export actions around to fix linter errors