Re-encode missing files + basic template source code export (#13)
Restart encoding of files reported as missing during export, and implement a basic template source code export option (without applying presentational and advanced settings, but linking in the LOOP and MAIN sequence).





* Re-encode missing files, stub out template export

Missing encoded item files are reported back through a custom error
class. The export actions check for this and trigger project actions
that remove encoded items, and begin analysing all files again. The user
has to wait until this new analysis is complete and then restart the
export.

The template export re-uses the audioWorker, giving it a path to an
audio directory in its own outputDir. It assumes that the template is
installed in node_modules and can be resolved - so far it must be placed
there manually, because it is not published on artifactory yet.

* Add bbcat-orchestration-template as a dependency to automatically install it.

* Wait for template files to copy, then modify config.js to link to sequences.

* Remove erroneously added dependency 'add'